debugging building under 64bit
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5284 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/guisim/GUIVehicleControl.cpp
@@ -89,10 +89,10 @@ GUIVehicleControl::getVehicleNames()
 }
 
 
-std::vector<size_t>
+std::vector<GLuint>
 GUIVehicleControl::getVehicleIDs()
 {
-    std::vector<size_t> ret;
+    std::vector<GLuint> ret;
     ret.reserve(myVehicleDict.size());
     for (VehicleDictType::iterator i=myVehicleDict.begin(); i!=myVehicleDict.end(); ++i) {
         MSVehicle *veh = (*i).second;
